feat(render): semantic board model + core encoder, move board types to core
Convert the board GUI surface to a semantic core render model and move its
domain types out of the editor protocol namespace:

- Minga.RenderModel.UI.Board + Board.Card replace the editor-owned
  Protocol.GUI.BoardPayload/BoardCardPayload and the pre-encoded board model.
- The core BoardEncoder owns the full wire format (card fields, flags, Float16
  sparklines, validation) as a pure function of the model and caches.
- BoardBuilder carries the shell's board payload through as the render model;
  AgentContextBuilder now consumes the core Board types, clearing it from the
  legacy builder allowlist too.
- Delete encode_gui_board and the legacy encode_gui_agent_context wrapper from
  ProtocolGUI; the wire-format test moves to BoardEncoderTest and the
  agent_context parity tests collapse onto golden-byte assertions.

Shrinks both render guardrail allowlists (board + board_builder +
agent_context_builder). On main no production code constructs a board payload
(Traditional.gui_payload/1 returns nil), so this is a pure type move; the board
shell extension will construct the core type when it lands. Wire format
unchanged.

Part of #2106.
