+++ lib/minga_editor/render_model/ui/agent_context_builder.ex
@@ -2,12 +2,11 @@ defmodule MingaEditor.RenderModel.UI.AgentContextBuilder do
   @moduledoc false
 
   alias Minga.RenderModel.UI.AgentContext
-  alias MingaEditor.Frontend.Protocol.GUI.BoardCardPayload
-  alias MingaEditor.Frontend.Protocol.GUI.BoardPayload
+  alias Minga.RenderModel.UI.Board
 
   @spec build(term()) :: AgentContext.t()
-  def build({:board, %BoardPayload{} = board}) do
-    card = BoardPayload.zoomed_card(board)
+  def build({:board, %Board{} = board}) do
+    card = Board.zoomed_card(board)
     build_from_card(board.zoomed_card_id, card)
   end
 
@@ -15,13 +14,13 @@ defmodule MingaEditor.RenderModel.UI.AgentContextBuilder do
     %AgentContext{visible: false}
   end
 
-  @spec build_from_card(pos_integer() | nil, BoardCardPayload.t() | nil) :: AgentContext.t()
+  @spec build_from_card(pos_integer() | nil, Board.Card.t() | nil) :: AgentContext.t()
   defp build_from_card(_card_id, nil) do
     %AgentContext{visible: false}
   end
 
-  defp build_from_card(_card_id, %BoardCardPayload{} = card) do
-    if BoardCardPayload.you_card?(card) do
+  defp build_from_card(_card_id, %Board.Card{} = card) do
+    if Board.Card.you_card?(card) do
       %AgentContext{visible: false}
     else
       %AgentContext{
